requirements.txt
pyyaml
raven
requests
beautifulsoup4
bcrypt
clint